Added boolean in constructor that decides whether to use the startPrice or not.
